Integrated indexv2 and refactor frontend
